fix(cleanup): offset overflow, constant-time code check, module docs
Three independent correctness/quality fixes from a codebase review, grouped into one cleanup PR.

- Pagination offset truncation: the contacts and tickets list services cast `pagination.offset() as i32`, so a large `?page=` (offset above i32::MAX) wrapped to a negative i32 and Postgres rejected the OFFSET, returning HTTP 500 on the list endpoints. Widen the five casts to `as i64` to match the ~30 other call sites; `limit` stays i32 because per_page is bounded.
- Non-constant-time hash comparison: `verify_login_approval` compared the login-approval code hashes with `==`. Switch to `constant_time_eq::constant_time_eq` over the byte slices, consistent with the TOTP, RMM webhook, and bunyip webhook secret checks.
- Stale module-status docs: `CLAUDE.md` and `docs/dev-docs/codebase-state.md` claimed only auth/contacts/tenants/tickets have real handlers and most route groups return HTTP 501. `api/router.rs` now nests/merges ~30 implemented modules and the `stub_routes()` 501 mechanism is gone; the sole remaining HTTP 501 is the PDF report-export format. Correct both docs.

#PMS-684

+++ src/modules/contacts/service.rs
@@ -321,7 +321,7 @@ impl ContactService {
         filter: &CompanyFilter,
         pagination: &PaginationParams,
     ) -> AppResult<(Vec<Company>, u64)> {
-        let offset = pagination.offset() as i32;
+        let offset = pagination.offset() as i64;
         let limit = pagination.limit() as i32;
 
         // Parallel WHERE clauses so the data and count queries each get
@@ -1026,7 +1026,7 @@ impl ContactService {
         filter: &ContactFilter,
         pagination: &PaginationParams,
     ) -> AppResult<(Vec<Contact>, u64)> {
-        let offset = pagination.offset() as i32;
+        let offset = pagination.offset() as i64;
         let limit = pagination.limit() as i32;
 
         // Parallel WHERE clauses so the data and count queries each get
@@ -1189,7 +1189,7 @@ impl ContactService {
         tenant_id: TenantId,
         pagination: &PaginationParams,
     ) -> AppResult<(Vec<CompanyIndustryResponse>, u64)> {
-        let offset = pagination.offset() as i32;
+        let offset = pagination.offset() as i64;
         let limit = pagination.limit() as i32;
         let mut tx = self.db.begin_with_tenant(tenant_id).await?;
         let total: i64 =
